Navigation on More Info Button, Formatted Movie Page CSS
Applied navigation of "More Info" button on homepage to redirect to movie, however, movie data in its page needs to be fetched and formatted properly before being able to replace textual template with dynamic information. Code's CSS is reformatted to reduce the amount of hardcode while trying to match the layout of original design.